refactor: unify FileSystemResolver into core with injected file discovery
The cli and vsx packages carried ~300-line near-identical copies of the
filesystem resolver, already diverging in their discovery strategy. The
shared implementation (index building, validation, load/cache,
resolve/getSync) now lives in core — which already uses fs for parse() —
with file discovery injected via a findFlowFiles option, so core gains no
glob dependency. cli and vsx are thin wrappers supplying their own glob
strategy. Drops the unused fromDirectory/fromFlow convenience statics.
